+++ Lectures/2020.02.04-EvenMoreGenerics/GenericStuff/GenericStuff/GenericStuff.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+</Project>
