+++ Lectures/2020.02.04-EvenMoreGenerics/GenericStuff/GenericStuff/IDescription.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GenericStuff
+{
+    public interface IDescription
+    {
+        string Description { get; }
+    }
+}
