--- Lectures/2020.02.04-EvenMoreGenerics_StartExpressions/GenericStuff/GenericStuff.Tests/SampleCanDoDescription.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace GenericStuff.Tests
-{
-    class SampleCanDoDescription : IDescription
-    {
-        public string Description => typeof(SampleCanDoDescription).ToString();
-    }
-}
