--- Lectures/2020.02.04-EvenMoreGenerics_StartExpressions/GenericStuff/GenericStuff.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericStuff", "GenericStuff\GenericStuff.csproj", "{1371AB7C-1091-43FA-9ABE-03029BB2E5EC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericStuff.Tests", "GenericStuff.Tests\GenericStuff.Tests.csproj", "{A8171A24-86A8-4150-96C2-5CEF28A05373}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1371AB7C-1091-43FA-9ABE-03029BB2E5EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1371AB7C-1091-43FA-9ABE-03029BB2E5EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1371AB7C-1091-43FA-9ABE-03029BB2E5EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1371AB7C-1091-43FA-9ABE-03029BB2E5EC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A8171A24-86A8-4150-96C2-5CEF28A05373}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8171A24-86A8-4150-96C2-5CEF28A05373}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8171A24-86A8-4150-96C2-5CEF28A05373}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8171A24-86A8-4150-96C2-5CEF28A05373}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9DB57AF4-B721-4CD9-8423-2BDFC7B61FFB}
-	EndGlobalSection
-EndGlobal
