--- Lectures/2020.02.04-EvenMoreGenerics_StartExpressions/GenericStuff/GenericStuff/IDescription.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace GenericStuff
-{
-    public interface IDescription
-    {
-        string Description { get; }
-    }
-}
